backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Task queue
celery[redis]==5.4.0
redis==5.2.0

# HTTP + scraping
httpx==0.28.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Tech detection
builtwith==1.3.5

# Web search
tavily-python==0.5.0

# AI providers
anthropic==0.40.0
openai==1.57.0

# PDF export
playwright==1.49.0

# Config + validation
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Utils
python-multipart==0.0.19
markdown==3.7

# Sync driver for Alembic + Celery DB access
psycopg2-binary==2.9.10
